backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0